Canonicalize and shrink the PackedState representation.
The identity of the active player is entirely determined by the
board position. If an even number of cards have been played, it's
Red's turn. If an odd number, it's Black's turn. This saves one byte
of PackedRepresentation.

The positions "3r 4b" and "4b 3r" are essentially the same position,
just mirrored horizontally. Use the same matchbox for both positions,
thus doubling the rate at which the matchbox player learns.